boolify the utilities.
This patch eliminates all uses of TRUE and FALSE that are strictly
local to the utilities, and replaces appropriate int variables with
bools. The giflib library code itself and its heder are not touched;
remaining uses of TRUE and FALSE are precisely those required to leave
the arguments to the library functions as ints.

The regression test passes.