GBA removal part 1.
The Game Boy Advanced was discontinued in 2008.  But the real kiss of death is
that *nobody documented this code*.  If you want it back in the distribution,
you know what to do!

Removing the substantial stuff as one commit, will remove hooks in the main code
with another one.  The repo history is cleaner that way.