Fix misdecoding when a GIF overflows the code table.
Version 4.1.6 and 4.2.1 do not decode this image correctly: http://ge.tt/15VJnNB/v/52

This is caused by the fact that DGifDecompressLine() overwrites the
last entry of the code table again and again in the case that the
table is full.

This issue was originally reported here:
https://github.com/muennich/sxiv/issues/75

All regression tests pass with this change.