Separate image header and allocation phases during reading,
this will allow users of giflib to inspect the properties of each saved
image without writing it to the SavedImages array. this allows a new
type of decoder that does frame-at-a-time without allocating memory for
all the frames in the gif simultaneously. this can yield large memory
allocation savings for gifs with large dimensions and many frames